+++ module-payment-service/src/main/java/com/shop/paymentservice/payment/entity/PaymentType.java
@@ -1,4 +1,4 @@
-package com.shop.paymentservice.entity;
+package com.shop.paymentservice.payment.entity;
 
 public enum PaymentType {
     PENDING,
