+++ module-payment-service/src/main/java/com/shop/paymentservice/payment/repository/PaymentRepository.java
@@ -1,6 +1,6 @@
-package com.shop.paymentservice.repository;
+package com.shop.paymentservice.payment.repository;
 
-import com.shop.paymentservice.entity.Payment;
+import com.shop.paymentservice.payment.entity.Payment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PaymentRepository extends JpaRepository<Payment, Long> {
